# public/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simulark API",
    "version": "1.0.0",
    "description": "API for Simulark Context Bridge and AI Orchestrator"
  },
  "paths": {
...
    "/api/generate": {
      "post": {
        "summary": "Generate Architecture",
        "description": "Predict and generate an architecture graph based on a natural language prompt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Natural language description of the architecture (e.g., 'Three-tier AWS app')"
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "AWS",
                      "GCP",
                      "Azure",
                      "Generic"
                    ],
                    "default": "Generic"
                  }
                },
                "required": [
                  "prompt"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated architecture",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    },
                    "plan": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or generation failure"
          }
        }
      }
    },
...
  }
}